--- Capstone_/Algorithm/Graphable.swift
@@ -1,17 +1,0 @@
-//
-//  Graphable.swift
-//  Capstone_
-//
-//  Created by Seokhyun Hong on 10/16/24.
-//
-
-protocol Graphable {
-  associatedtype Element: Hashable // 1
-  var description: CustomStringConvertible { get } // 2
-  
-  func createVertex(data: Element) -> Vertex<Element> // 3
-  func add(_ type: EdgeType, from source: Vertex<Element>, to destination: Vertex<Element>, weight: Double?) // 4
-  func weight(from source: Vertex<Element>, to destination: Vertex<Element>) -> Double? // 5
-  func edges(from source: Vertex<Element>) -> [Edge<Element>]? // 6
-}
-
