+++ Capstone_/Datas/NodesData.swift
@@ -132,3 +132,126 @@ let mapNodes: [nodes] = [
     nodes(name: "59", x: -235.04113, y: -212.58353, z: 1)
 ]
 
+let mapNodesTwo: [nodes] = [
+    // Classroom Nodes
+    nodes(name: "Restroom 2010", x: 6.385, y: -0.503, z: 0.2),
+    nodes(name: "Stair C2001", x: 5.607, y: -0.314, z: 0.2),
+    nodes(name: "Stair C2012", x: -6.177, y: 4.757, z: 0.2),
+    nodes(name: "Stair C2017", x: 6.414, y: 5.796, z: 0.2),
+    nodes(name: "Room 2020", x: 4.375, y: -0.583, z: 0.2),
+    nodes(name: "Room 2030", x: 4.676, y: 0.21, z: 0.2),
+    nodes(name: "Room 2040", x: 4.382, y: 1.718, z: 0.2),
+    nodes(name: "Room 2050", x: 3.954, y: 1.876, z: 0.2),
+    nodes(name: "Room 2060", x: 3.853, y: 0.516, z: 0.2),
+    nodes(name: "Room 2100", x: 3.368, y: -2.10, z: 0.2),
+    nodes(name: "Room 2150", x: 2.833, y: -4.396, z: 0.2),
+    nodes(name: "Room 2160", x: 2.58, y: -0.928, z: 0.2),
+    nodes(name: "Restroom 2170", x: 1.883, y: -0.80, z: 0.2),
+    nodes(name: "Room 2200", x: 1.07, y: 1.10, z: 0.2),
+    nodes(name: "Room 2210", x: 1.07, y: 1.365, z: 0.2),
+    nodes(name: "Room 2220", x: 1.07, y: 3.719, z: 0.2),
+    nodes(name: "Room 2230", x: 1.07, y: 3.981, z: 0.2),
+    nodes(name: "Room 2240", x: -1.023, y: 3.978, z: 0.2),
+    nodes(name: "Room 2250", x: -1.023, y: 3.722, z: 0.2),
+    nodes(name: "Room 2260", x: -1.023, y: 1.365, z: 0.2),
+    nodes(name: "Room 2270", x: -1.023, y: 1.104, z: 0.2),
+    nodes(name: "Room 2280", x: -1.70, y: -0.613, z: 0.2),
+    nodes(name: "Room 2285", x: -2.31, y: -0.629, z: 0.2),
+    nodes(name: "Room 2290", x: -3.885, y: -0.455, z: 0.2),
+    nodes(name: "Room 2295", x: -4.134, y: -0.455, z: 0.2),
+    nodes(name: "Room 2300", x: -4.117, y: -2.582, z: 0.2),
+    nodes(name: "Room 2310", x: -3.863, y: -3.768, z: 0.2),
+    nodes(name: "Room 2320", x: -2.787, y: -4.396, z: 0.2),
+    nodes(name: "Room 2330a", x: -5.085, y: -3.268, z: 0.2),
+    nodes(name: "Room 2330b", x: -4.288, y: -4.253, z: 0.2),
+    nodes(name: "Room 2340", x: -5.598, y: -2.416, z: 0.2),
+    nodes(name: "Room 2350", x: -5.417, y: -0.481, z: 0.2),
+    nodes(name: "Restroom 2360", x: -6.447, y: 0.10, z: 0.2),
+    nodes(name: "Restroom 2370", x: -6.75, y: 0.497, z: 0.2),
+    nodes(name: "Room 2400A", x: -5.095, y: 2.12, z: 0.2),
+    nodes(name: "Room 2400B", x: -4.845, y: 2.12, z: 0.2),
+    nodes(name: "Room 2400C", x: -4.227, y: 1.91, z: 0.2),
+    nodes(name: "Room 2400D", x: -4.025, y: 2.12, z: 0.2),
+    nodes(name: "Room 2400E", x: -3.775, y: 2.12, z: 0.2),
+    nodes(name: "Room 2400F", x: -3.36, y: 2.005, z: 0.2),
+    nodes(name: "Room 2400G", x: -3.36, y: 1.755, z: 0.2),
+    nodes(name: "Room 2400H", x: -3.36, y: 0.965, z: 0.2),
+    nodes(name: "Room 2400J", x: -3.36, y: 0.71, z: 0.2),
+    nodes(name: "Room 2400K", x: -3.775, y: 0.60, z: 0.2),
+    nodes(name: "Room 2400L", x: -4.025, y: 0.60, z: 0.2),
+    nodes(name: "Room 2400M", x: -4.22, y: 0.81, z: 0.2),
+    nodes(name: "Room 2400N", x: -4.845, y: 0.60, z: 0.2),
+    nodes(name: "Room 2400P", x: -5.095, y: 0.60, z: 0.2),
+    nodes(name: "Room 2420", x: -6.01, y: 0.916, z: 0.2),
+    nodes(name: "Room 2430", x: -6.01, y: 2.287, z: 0.2),
+    nodes(name: "Room 2440", x: -6.30, y: 3.766, z: 0.2),
+    nodes(name: "Room 2460", x: -6.45, y: 5.095, z: 0.2),
+    nodes(name: "Room 2480", x: -6.171, y: 5.532, z: 0.2),
+    nodes(name: "Room 2500A", x: -5.098, y: 4.50, z: 0.2),
+    nodes(name: "Room 2500B", x: -4.849, y: 4.50, z: 0.2),
+    nodes(name: "Room 2500C", x: -4.03, y: 4.50, z: 0.2),
+    nodes(name: "Room 2500D", x: -3.777, y: 4.50, z: 0.2),
+    nodes(name: "Room 2500E", x: -3.35, y: 4.301, z: 0.2),
+    nodes(name: "Room 2500F", x: -3.35, y: 4.04, z: 0.2),
+    nodes(name: "Room 2500G", x: -3.229, y: 3.40, z: 0.2),
+    nodes(name: "Room 2500H", x: -3.779, y: 3.40, z: 0.2),
+    nodes(name: "Room 2500J", x: -4.029, y: 3.40, z: 0.2),
+    nodes(name: "Room 2500M", x: -4.839, y: 3.40, z: 0.2),
+    nodes(name: "Room 2500N", x: -5.097, y: 3.40, z: 0.2),
+    nodes(name: "Room 2500K", x: -4.315, y: 3.68, z: 0.2),
+    nodes(name: "Room 2500L", x: -4.566, y: 3.68, z: 0.2),
+    nodes(name: "Room 2620", x: -5.89, y: 5.532, z: 0.2),
+    nodes(name: "Room 2640", x: -4.539, y: 5.532, z: 0.2),
+    nodes(name: "Room 2660", x: -3.199, y: 5.532, z: 0.2),
+    nodes(name: "Room 2680", x: -1.857, y: 5.532, z: 0.2),
+    nodes(name: "Room 2820", x: 1.906, y: 5.531, z: 0.2),
+    nodes(name: "Room 2840", x: 3.249, y: 5.531, z: 0.2),
+    nodes(name: "Room 2860", x: 4.594, y: 5.531, z: 0.2),
+    nodes(name: "Room 2880", x: 5.868, y: 5.531, z: 0.2),
+    nodes(name: "Room 2910", x: 4.336, y: 5.127, z: 0.2),
+    nodes(name: "Room 2920", x: 4.641, y: 5.009, z: 0.2),
+
+
+
+
+
+    nodes(name: "1", x: 0.324, y: -5.262, z: 0.2),
+    nodes(name: "2", x: 0.324, y: -4.034, z: 0.2),
+    nodes(name: "3", x: 2.90, y: -4.034, z: 0.2),
+    nodes(name: "4", x: 2.90, y: -0.245, z: 0.2),
+    nodes(name: "5", x: 4.25, y: -0.084, z: 0.2),
+    nodes(name: "6", x: 1.912, y: -0.245, z: 0.2),
+    nodes(name: "7", x: 0.90, y: -0.245, z: 0.2),
+    nodes(name: "8", x: 0.90, y: 5.30, z: 0.2),
+    nodes(name: "10", x: 6.20, y: 5.40, z: 0.2),
+    nodes(name: "11", x: 6.00, y: 0.00, z: 0.2),
+    nodes(name: "12", x: 6.582, y: 0.07, z: 0.2),
+    nodes(name: "13", x: -0.267, y: -5.262, z: 0.2),
+    nodes(name: "14", x: -0.267, y: -4.028, z: 0.2),
+    nodes(name: "15", x: -4.052, y: -4.028, z: 0.2),
+    nodes(name: "16", x: -4.24, y: -3.284, z: 0.2),
+    nodes(name: "17", x: -4.80, y: -2.80, z: 0.2),
+    nodes(name: "18", x: -4.953, y: -2.343, z: 0.2),
+    nodes(name: "19", x: -4.953, y: -1.308, z: 0.2),
+    nodes(name: "20", x: -4.953, y: -0.272, z: 0.2),
+    nodes(name: "21", x: -2.458, y: -0.272, z: 0.2),
+    nodes(name: "22", x: -1.84, y: -0.272, z: 0.2),
+    nodes(name: "23", x: -0.831, y: -0.272, z: 0.2),
+    nodes(name: "24", x: -0.80, y: 5.30, z: 0.2),
+    nodes(name: "25", x: -5.746, y: 5.30, z: 0.2),
+    nodes(name: "26", x: -5.746, y: 4.83, z: 0.2),
+    nodes(name: "27", x: -5.746, y: 4.36, z: 0.2),
+    nodes(name: "28", x: -3.60, y: 4.36, z: 0.2),
+    nodes(name: "29", x: -3.60, y: 3.574, z: 0.2),
+    nodes(name: "30", x: -5.746, y: 3.574, z: 0.2),
+    nodes(name: "31", x: -5.746, y: 2.787, z: 0.2),
+    nodes(name: "32", x: -5.746, y: 2.00, z: 0.2),
+    nodes(name: "33", x: -4.648, y: 2.00, z: 0.2),
+    nodes(name: "34", x: -3.541, y: 2.00, z: 0.2),
+    nodes(name: "35", x: -3.541, y: 1.361, z: 0.2),
+    nodes(name: "36", x: -3.541, y: 0.722, z: 0.2),
+    nodes(name: "37", x: -4.648, y: 0.722, z: 0.2),
+    nodes(name: "38", x: -5.755, y: 0.722, z: 0.2),
+    nodes(name: "39", x: -5.755, y: -0.269, z: 0.2),
+    nodes(name: "40", x: 2.923, y: -2.117, z: 0.2)
+]
