+++ Capstone_/Views/ContentView.swift
@@ -4,7 +4,6 @@ import SwiftUI
 struct ContentView: View {
     @State private var isSplashActive = false
     @EnvironmentObject var mapViewModel: MapViewModel
-    @EnvironmentObject var mapViewModel2: MapViewModel2
     @EnvironmentObject var arViewModel: ARViewModel
 
     var body: some View {
@@ -14,8 +13,8 @@ struct ContentView: View {
                    
 //                    CampusNavigatorView()
                     
-//                    ARView(viewModel: arViewModel)
-//                        .frame(maxWidth: .infinity, maxHeight: UIScreen.main.bounds.height * 0.7)
+                    ARView(viewModel: arViewModel)
+                        .frame(maxWidth: .infinity, maxHeight: UIScreen.main.bounds.height * 0.7)
                     
                     MapView(viewModel: mapViewModel) // navigate to MapView floor 1
                         .frame(maxWidth: .infinity, maxHeight: UIScreen.main.bounds.height * 0.3)
